SOH-843: Actualización DTOs y ui para Detalles de movimientos (Accounts, Cards & Sent Transfers)

+++ lib/application/daily_banking/accounts/wires/sent_transfers/simplified/simplified_sent_transfers_controller.dart
@@ -56,8 +56,8 @@ class SimplifiedSentTransfersController extends StateNotifier<SimplifiedSentTran
   }
 
   Future<void> updateFilter({
-    required int? amountFrom,
-    required int? amountTo,
+    required double? amountFrom,
+    required double? amountTo,
     required DateTime? dateFrom,
     required DateTime? dateTo,
   }) async {
